move param utils to params
