+++ neurodsp/sim/utils.py
@@ -163,69 +163,3 @@ def modulate_signal(sig, modulation, fs=None, mod_params=None):
     msig = sig * modulation
 
     return msig
-
-## Utilities for helping with parameter management
-
-BASE_PARAMS = ['n_seconds', 'fs']
-
-def get_base_params(params):
-    """Get base parameters from a parameter definition.
-
-    Parameters
-    ----------
-    params : dict
-        Parameter definition.
-
-    Returns
-    -------
-    params : dict
-        Base parameters.
-    """
-
-    return {key : value for key, value in params.items() if key in BASE_PARAMS}
-
-
-def drop_base_params(params):
-    """Drop base parameters from a parameter definition.
-
-    Parameters
-    ----------
-    params : dict
-        Parameter definition.
-
-    Returns
-    -------
-    params : dict
-        Parameter definition, excluding base parameters.
-    """
-
-    return {key : value for key, value in params.items() if key not in BASE_PARAMS}
-
-
-def get_param_values(params, extract=None, component=None):
-    """Get a set of parameter values from a set of parameter definitions.
-
-    Parameters
-    ----------
-    params : list of dict
-        Parameter definitions for multiple simulations.
-    extract : str
-        Name of the parameter to extract.
-    component : str, optional
-        Which component to extract the parameter from.
-        Only used if the parameter definition is for a multi-component simulation.
-
-    Returns
-    -------
-    values : list
-        Extracted parameter values.
-    """
-
-    if component:
-        values =[cparams['components'][component][extract] for cparams in params]
-    elif extract:
-        values = [cparams[extract] for cparams in params]
-    else:
-        values = None
-
-    return values
