+++ neurodsp/tests/sim/test_utils.py
@@ -34,32 +34,3 @@ def test_modulate_signal(tsig):
     # Check modulation passing in a 1d array directly
     msig2 = modulate_signal(tsig, tsig)
     check_sim_output(msig2)
-
-def test_get_base_params():
-
-    params = {'n_seconds' : 2, 'fs' : 250, 'exponent' : -1}
-    out1 = get_base_params(params)
-    for bparam in out1:
-        assert bparam in BASE_PARAMS
-
-def test_drop_base_params():
-
-    params = {'n_seconds' : 2, 'fs' : 250, 'exponent' : -1}
-    out1 = drop_base_params(params)
-    for bparam in BASE_PARAMS:
-        assert bparam not in out1
-    assert 'exponent' in params
-
-def test_get_param_values():
-
-    params = [{'n_seconds' : 2, 'fs' : 250, 'exponent' : -2},
-              {'n_seconds' : 2, 'fs' : 250, 'exponent' : -1}]
-    assert get_param_values(params, 'exponent') == [-2, -1]
-    assert get_param_values(params, 'n_seconds') == [2, 2]
-
-    params = [{'n_seconds' : 2, 'fs' : 250, 'components' : \
-                {'sim_powerlaw' : {'exponent' : -2}, 'sim_oscillation' : {'freq' : 10}}},
-              {'n_seconds' : 2, 'fs' : 250, 'components' : \
-                {'sim_powerlaw' : {'exponent' : -1}, 'sim_oscillation' : {'freq' : 10}}}]
-    assert get_param_values(params, 'exponent', 'sim_powerlaw') == [-2, -1]
-    assert get_param_values(params, 'freq', 'sim_oscillation') == [10, 10]
